csv-ai-final/pyproject.toml
[project]
name = "csv-analysis"
version = "0.1.0"
description = "CSV analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "requests>=2.31.0",
    "pandas>=2.3.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "streamlit>=1.28.0",
]
